🚚 rename: App.js app 폴더 밖으로 이동
